added delete-components;

+++ src/tasks/delete-component.js
@@ -5,9 +5,10 @@ const { getComponentsFromName } = require('./migrations/utils')
  *
  * @param api {Object}
  * @param comp {String | Number}
+ * @param dryrun {Boolean}
  * @returns {Promise<void>}
  */
-const deleteComponent = async (api, { comp }) => {
+const deleteComponent = async (api, { comp, dryrun = false }) => {
   try {
     let component
     if (!isNaN(comp)) {
@@ -19,8 +20,10 @@ const deleteComponent = async (api, { comp }) => {
     if (Object.keys(component).length === 0) {
       return Promise.reject(new Error(`Component ${comp} not found.`))
     }
-    await api.delete(`components/${component.id}`)
-    console.log(chalk.green('✓') + ` Component '${component.name}' deleted.`)
+    if (!dryrun) {
+      await api.delete(`components/${component.id}`)
+    }
+    console.log(chalk.green('✓') + ' Component ' + chalk.blue(component.name) + ' deleted.')
   } catch (e) {
     console.error(`${chalk.red('X')} An error occurred in delete-component task.`)
     return Promise.reject(new Error(e))
